+++ tests/test_variance_box.py
@@ -30,7 +30,7 @@ def test_variance_box():
         )
     )
 
-    result = cle.create(test1)
+    result = cle.create(test1, dtype=float)
     cle.variance_box(test1, result, 1, 1, 0)
 
     a = cle.pull(result)
